requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
passlib
bcrypt
python-jose
pymongo
requests
httpx
google-generativeai
python-dotenv
groq
openai
deepgram-sdk
websockets
fpdf
matplotlib
pandas